requirements.txt
absl-py==0.6.1
astor==0.7.1
azure-functions==1.0.0a4
azure-functions-worker==1.0.0a4
certifi==2018.10.15
chardet==3.0.4
gast==0.2.0
grpcio==1.14.2
grpcio-tools==1.14.2
h5py==2.8.0
idna==2.7
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
numpy==1.15.4
pandas==0.23.4
Pillow==5.3.0
protobuf==3.6.1
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
requests==2.20.1
scipy==1.1.0
six==1.11.0
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
urllib3==1.24.1
Werkzeug==0.14.1
